split out server routes into separate files, more with the chat and federation stuff
