+++ Documentation/Federation.tid
@@ -1,0 +1,89 @@
+title: $:/plugins/OokTech/Bob/Federation
+caption: Federation
+tags: [[Bob Documentation]]
+description: Federation between servers
+
+!! Federation
+
+Federation lets you communicate between different servers.
+
+This allows many things, including:
+
+* syncing wikis
+* fetching tiddlers from wikis on remote servers
+* sending tiddlers to wikis on remote servers
+* real-time chat between different servers
+
+!! Backend things
+
+The local server saves information about remote servers. This includes:
+
+```
+uniqueName: {
+  serverName: human readable name,
+  url: the most recent url used to connect to the remote server,
+  staticUrl: if the url changes or not (yes or no),
+  publicKey: a public key,
+  availableWikis: {
+    wikiName1: {
+      sync: yes/no,
+      syncFilter: <<someFilter>>,
+      syncType: pull,push,remoteWins,localWins,newestWins,
+      autoSync: yes/no,
+      public: yes/no
+    },
+    wikiName2: {
+      <<same stuff>>
+    }
+  },
+  chats: {
+    chatName1: {
+      public: yes/no,
+      relay: yes/no
+    }
+  }
+}
+```
+
+to connect to another server select the serverName from a select widget,
+then there is an interface that has the availableWikis and the current sync
+status and setup for the wiki.
+
+Servers have to have a login interface as well so you can log in to see private
+wikis and edit wikis.
+
+For this the login is done using either username/password pairs or by a signed
+token that the remote server can check against the public key it has listed.
+This should probably use this process:
+
+# Local server requests a login to the remote wiki
+# Remote server replies with a random number
+# Local server puts the random number into a signed token and sends it back to the remote server, if everything matches than this is the access token for further interactions.
+# Because the token handshake is taken care of by the servers without needing human interaction the tokens can be changed as often as desired, possibly using a unique random number for each token sent.
+
+To connect to a new server there is an interface where you enter the URL and it
+requests information from the remote server.
+The requested information is:
+
+* human readable server name
+* public key
+* available wikis
+* if the server hosts chats
+** if so which chats are currently available
+** if new chats can be created using this server
+
+this information is saved locally and can be used in the above interface.
+
+!!! Chat
+
+each server can host a chat server as well.
+Each chat on a server has a unique name and can be public or private.
+
+When you connect to a server you get a list of chats available on that server.
+
+then in the connection interface above you select the server and then a list of
+chats that are available on that server appear, you can select a chat and the
+chat interface for that chat appears and you get an update of the chat history
+for that server.
+
+You probably have to click a button to load the conversation.
