+++ commands/ws-server.js
@@ -37,12 +37,9 @@ if($tw.node) {
     return false;
   }
 
-  const util = require("util"),
-    fs = require("fs"),
-    url = require("url"),
+  const url = require("url"),
     path = require("path"),
-    http = require("http"),
-    qs = require("querystring");
+    http = require("http")
 
   /*
   A simple HTTP server with regexp-based routes
@@ -268,460 +265,9 @@ if($tw.node) {
       wiki: this.commander.wiki
     });
     // Add route handlers
-    $tw.httpServer.addRoute({
-      method: "GET",
-      path: /^\/$/,
-      handler: function(request,response,state) {
-        const token = $tw.Bob.getCookie(request.headers.cookie, 'token');
-        const authorised = $tw.Bob.AccessCheck('RootWiki', token, 'view');
-        if(authorised) {
-          let text;
-          // Load the wiki
-          const exists = $tw.ServerSide.loadWiki('RootWiki');
-          if(exists) {
-            // Get the raw html to send
-            text = $tw.ServerSide.prepareWiki('RootWiki', true);
-          } else {
-            text = "<html><p>RootWiki not found! If you have autoUnloadWikis set to true setting it to false may fix this problem.</p></html>"
-          }
-          // Send the html to the server
-          response.writeHead(200, {"Content-Type": state.server.get("serveType")});
-          response.end(text,"utf8");
-        }
-      }
+    $tw.modules.forEachModuleOfType("serverroute", function(title, routeDefinition) {
+      $tw.httpServer.addRoute(routeDefinition);
     });
-    // Add favicon route
-    $tw.httpServer.addRoute({
-      method: "GET",
-      path: /^\/favicon.ico$/,
-      handler: function(request,response,state) {
-        const token = $tw.Bob.getCookie(request.headers.cookie, 'token');
-        const authorised = $tw.Bob.AccessCheck('RootWiki', token, 'view');
-        if(authorised) {
-          // Load the wiki
-          const exists = $tw.ServerSide.loadWiki('RootWiki');
-          let buffer = ''
-          if(exists) {
-            response.writeHead(200, {"Content-Type": "image/x-icon"});
-            if($tw.Bob.Wikis['RootWiki']) {
-              buffer = $tw.Bob.Wikis['RootWiki'].wiki.getTiddlerText('$:/favicon.ico')
-            }
-          } else {
-            buffer = "";
-          }
-          response.end(buffer,"base64");
-        } else {
-          response.writeHead(404);
-          response.end();
-        }
-      }
-    });
-    $tw.settings.API = $tw.settings.API || {};
-    if($tw.settings.API.pluginLibrary === 'yes') {
-      const getPluginList = function () {
-        let pluginList = []
-        if(typeof $tw.settings.pluginsPath === 'string') {
-          const basePath = $tw.ServerSide.getBasePath();
-          const pluginsPath = path.resolve(basePath, $tw.settings.pluginsPath)
-          if(fs.existsSync(pluginsPath)) {
-            try {
-              const pluginAuthors = fs.readdirSync(pluginsPath)
-              pluginAuthors.forEach(function (author) {
-                const pluginAuthorPath = path.join(pluginsPath, './', author)
-                if(fs.statSync(pluginAuthorPath).isDirectory()) {
-                  const pluginAuthorFolders = fs.readdirSync(pluginAuthorPath)
-                  for(let t=0; t<pluginAuthorFolders.length; t++) {
-                    const fullPluginFolder = path.join(pluginAuthorPath,pluginAuthorFolders[t])
-                    const pluginFields = $tw.loadPluginFolder(fullPluginFolder)
-                    if(pluginFields) {
-                      let readme = ""
-                      let readmeText = ''
-                      try {
-                        // Try pulling out the plugin readme
-                        const pluginJSON = JSON.parse(pluginFields.text).tiddlers
-                        readme = pluginJSON[Object.keys(pluginJSON).filter(function(title) {
-                          return title.toLowerCase().endsWith('/readme')
-                        })[0]]
-                      } catch (e) {
-                        $tw.Bob.logger.error('Error parsing plugin', e, {level:1})
-                      }
-                      if(readme) {
-                        readmeText = readme.text
-                      }
-                      const nameParts = pluginFields.title.split('/')
-                      const name = nameParts[nameParts.length-2] + '/' + nameParts[nameParts.length-1]
-                      const listInfo = {
-                        name: name,
-                        description: pluginFields.description,
-                        tiddlerName: pluginFields.title,
-                        version: pluginFields.version,
-                        author: pluginFields.author,
-                        readme: readmeText
-                      }
-                      pluginList.push(listInfo)
-                    }
-                  }
-                }
-              })
-            } catch (e) {
-              $tw.Bob.logger.error('Problem loading plugin', e, {level:1})
-            }
-          }
-        }
-        return pluginList
-      }
-      const getPlugin = function (request) {
-        const urlParts = request.url.split('/')
-        if(typeof $tw.settings.pluginsPath === 'string') {
-          const basePath = $tw.ServerSide.getBasePath();
-          const pluginsPath = path.resolve(basePath, $tw.settings.pluginsPath)
-          const pluginPath = path.resolve(pluginsPath, urlParts[urlParts.length-2], urlParts[urlParts.length-1])
-          if(fs.statSync(pluginPath).isDirectory()) {
-            const pluginFields = $tw.loadPluginFolder(pluginPath)
-            return pluginFields
-          }
-        }
-        return false
-      }
-      // Add list route
-      const pluginListRoute = new RegExp('^\/api\/plugins\/list')
-      $tw.httpServer.addRoute({
-        method: "POST",
-        path: pluginListRoute,
-        handler: function (request, response, state) {
-          const token = $tw.Bob.getCookie(request.headers.cookie, 'token');
-          const authorised = $tw.Bob.AccessCheck("RootWiki", token, 'list');
-          if(authorised) {
-            const pluginList = getPluginList()
-            response.setHeader('Access-Control-Allow-Origin', '*')
-            response.writeHead(200)
-            response.end(JSON.stringify(pluginList))
-          } else {
-            response.writeHead(403)
-            response.end()
-          }
-        }
-      })
-      // Add plugin fetch route
-      const fetchPluginRoute = new RegExp('^\/api\/plugins\/fetch\/.+')
-      $tw.httpServer.addRoute({
-        method: "POST",
-        path: fetchPluginRoute,
-        handler: function (request, response, state) {
-          const token = $tw.Bob.getCookie(request.headers.cookie, 'token');
-          const authorised = $tw.Bob.AccessCheck("RootWiki", token, 'fetchPlugin');
-          if(authorised) {
-            const plugin = getPlugin(request)
-            if(plugin) {
-              response.setHeader('Access-Control-Allow-Origin', '*')
-              response.writeHead(200)
-              response.end(JSON.stringify(plugin))
-            } else {
-              response.writeHead(403)
-              response.end()
-            }
-          } else {
-            response.writeHead(403)
-            response.end()
-          }
-        }
-      })
-    }
-    if($tw.settings.API.enablePush === 'yes') {
-      const pushPathRegExp = new RegExp('^\/api\/push');
-      $tw.httpServer.addRoute({
-        method: "POST",
-        path: pushPathRegExp,
-        handler: function (request, response, state) {
-          let body = ''
-          request.on('data', function(chunk){
-            body += chunk;
-            // We limit the size of a push to 5mb for now.
-            if(body.length > 5e6) {
-              response.writeHead(413, {'Content-Type': 'text/plain'}).end();
-              request.connection.destroy();
-            }
-          });
-          request.on('end', function() {
-            try {
-              let bodyData = JSON.parse(body)
-              // Make sure that the token sent here matches the https header
-              // and that the token has push access to the toWiki
-              const token = $tw.Bob.getCookie(request.headers.cookie, 'token');
-              const authorised = $tw.Bob.AccessCheck(bodyData.toWiki, token, 'push');
-              if(authorised) {
-                if($tw.settings.wikis[bodyData.toWiki] || bodyData.toWiki === 'RootWiki') {
-                  $tw.ServerSide.loadWiki(bodyData.toWiki);
-                  // Make sure that the wiki exists and is loaded
-                  if($tw.Bob.Wikis[bodyData.toWiki]) {
-                    if($tw.Bob.Wikis[bodyData.toWiki].State === 'loaded') {
-                      if(bodyData.tiddlers && bodyData.toWiki) {
-                        Object.keys(bodyData.tiddlers).forEach(function(title) {
-                          bodyData.tiddlers[title].fields.modified = $tw.utils.stringifyDate(new Date(bodyData.tiddlers[title].fields.modified));
-                          bodyData.tiddlers[title].fields.created = $tw.utils.stringifyDate(new Date(bodyData.tiddlers[title].fields.created));
-                          $tw.syncadaptor.saveTiddler(bodyData.tiddlers[title], bodyData.toWiki);
-                        });
-                        response.writeHead(200)
-                        response.end()
-                      }
-                    }
-                  }
-                }
-              } else {
-                response.writeHead(400)
-                response.end()
-              }
-            } catch (e) {
-              response.writeHead(400)
-              response.end()
-            }
-          })
-        }
-      });
-    }
-    if($tw.settings.API.enableFetch === 'yes') {
-      const fetchPathRegExp = new RegExp('^\/api\/fetch&');
-      $tw.httpServer.addRoute({
-        method: "POST",
-        path: fetchPathRegExp,
-        handler: function(request,response,state) {
-          let body = ''
-          let list = []
-          let data = {}
-          response.setHeader('Access-Control-Allow-Origin', '*')
-          response.writeHead(200, {"Content-Type": "application/json"})
-          request.on('data', function(chunk){
-            body += chunk;
-            // We limit this to 1mb, it should never be anywhere near that
-            // big
-            if(body.length > 1e6) {
-              response.writeHead(413, {'Content-Type': 'text/plain'}).end();
-              request.connection.destroy();
-            }
-          });
-          request.on('end', function() {
-            body = body.replace(/^message=/, '')
-            try {
-              const bodyData = JSON.parse(body)
-              if(bodyData.filter && bodyData.fromWiki) {
-                // Make sure that the person has access to the wiki
-                const token = $tw.Bob.getCookie(request.headers.cookie, 'token');
-                const authorised = $tw.Bob.AccessCheck(bodyData.fromWiki, token, 'view');
-                if(authorised) {
-                  // Make sure that the wiki is listed
-                  if($tw.settings.wikis[bodyData.fromWiki] || bodyData.fromWiki === 'RootWiki') {
-                    // If the wiki isn't loaded than load it
-                    if(!$tw.Bob.Wikis[bodyData.fromWiki]) {
-                      $tw.ServerSide.loadWiki(bodyData.fromWiki);
-                    } else if($tw.Bob.Wikis[bodyData.fromWiki].State !== 'loaded') {
-                      $tw.ServerSide.loadWiki(bodyData.fromWiki);
-                    }
-                    // Make sure that the wiki exists and is loaded
-                    if($tw.Bob.Wikis[bodyData.fromWiki]) {
-                      if($tw.Bob.Wikis[bodyData.fromWiki].State === 'loaded') {
-                        list = $tw.Bob.Wikis[bodyData.fromWiki].wiki.filterTiddlers(bodyData.filter);
-                      }
-                    }
-                  }
-                }
-                let tiddlers = {};
-                let info = {};
-                list.forEach(function(title) {
-                  const tempTid = $tw.Bob.Wikis[bodyData.fromWiki].wiki.getTiddler(title);
-                  tiddlers[title] = tempTid;
-                  info[title] = {};
-                  if(bodyData.fieldList) {
-                    bodyData.fieldList.split(' ').forEach(function(field) {
-                      info[title][field] = tempTid.fields[field];
-                    })
-                  } else {
-                    info[title]['modified'] = tempTid.fields.modified;
-                  }
-                })
-                // Send the tiddlers
-                data = {list: list, tiddlers: tiddlers, info: info};
-                data = JSON.stringify(data) || "";
-                response.end(data);
-              }
-            } catch (e) {
-              data = JSON.stringify(data) || "";
-              response.end(data);
-            }
-          })
-        }
-      });
-      const fetchListPathRegExp = new RegExp('^\/api\/fetch\/list');
-      $tw.httpServer.addRoute({
-        method: "POST",
-        path: fetchListPathRegExp,
-        handler: function(request,response,state) {
-          let body = ''
-          let list
-          let data = {}
-          response.setHeader('Access-Control-Allow-Origin', '*')
-          response.writeHead(200, {"Content-Type": "application/json"})
-          request.on('data', function(chunk){
-            body += chunk;
-            // We limit this to 1mb, it should never be anywhere near that
-            // big
-            if(body.length > 1e6) {
-              response.writeHead(413, {'Content-Type': 'text/plain'}).end();
-              request.connection.destroy();
-            }
-          });
-          request.on('end', function() {
-            body = body.replace(/^message=/, '')
-            try {
-              const bodyData = JSON.parse(body)
-              if(bodyData.filter && bodyData.fromWiki) {
-                // Make sure that the person has access to the wiki
-                const token = $tw.Bob.getCookie(request.headers.cookie, 'token');
-                const authorised = $tw.Bob.AccessCheck(bodyData.fromWiki, token, 'view');
-                if(authorised) {
-                  // Make sure that the wiki is listed
-                  if($tw.settings.wikis[bodyData.fromWiki] || bodyData.fromWiki === 'RootWiki') {
-                    // If the wiki isn't loaded than load it
-                    if(!$tw.Bob.Wikis[bodyData.fromWiki]) {
-                      $tw.ServerSide.loadWiki(bodyData.fromWiki);
-                    } else if($tw.Bob.Wikis[bodyData.fromWiki].State !== 'loaded') {
-                      $tw.ServerSide.loadWiki(bodyData.fromWiki);
-                    }
-                    // Make sure that the wiki exists and is loaded
-                    if($tw.Bob.Wikis[bodyData.fromWiki]) {
-                      if($tw.Bob.Wikis[bodyData.fromWiki].State === 'loaded') {
-                        // Make a temp wiki to run the filter on
-                        let tempWiki = new $tw.Wiki();
-                        $tw.Bob.Wikis[bodyData.fromWiki].tiddlers.forEach(function(internalTitle) {
-                          const tiddler = $tw.wiki.getTiddler(internalTitle);
-                          let newTiddler = JSON.parse(JSON.stringify(tiddler));
-                          newTiddler.fields.modified = $tw.utils.stringifyDate(new Date(newTiddler.fields.modified));
-                          newTiddler.fields.created = $tw.utils.stringifyDate(new Date(newTiddler.fields.created));
-                          newTiddler.fields.title = newTiddler.fields.title.replace('{' + bodyData.fromWiki + '}', '');
-                          // Add all the tiddlers that belong in wiki
-                          tempWiki.addTiddler(new $tw.Tiddler(newTiddler.fields));
-                        })
-                        // Use the filter
-                        list = tempWiki.filterTiddlers(bodyData.filter);
-                      }
-                    }
-                  }
-                }
-                let tiddlers = {}
-                let info = {}
-                list.forEach(function(title) {
-                  const tempTid = tempWiki.getTiddler(title)
-                  info[title] = {}
-                  if(bodyData.fieldList) {
-                    bodyData.fieldList.split(' ').forEach(function(field) {
-                      info[title][field] = tempTid.fields[field];
-                    })
-                  } else {
-                    info[title]['modified'] = tempTid.fields.modified;
-                  }
-                })
-                // Send the info
-                data = {list: list, info: info}
-                data = JSON.stringify(data) || "";
-                response.end(data);
-              }
-            } catch (e) {
-              data = JSON.stringify(data) || "";
-              response.end(data);
-            }
-          })
-        }
-      });
-    }
-    if($tw.settings.enableFileServer === 'true') {
-      // Start with the same base path as the --listen command
-      let pathRegExp = new RegExp('\/files\/.+$');
-      if(typeof $tw.settings.fileURLPrefix === 'string' && ($tw.settings.fileURLPrefix !== '' || $tw.settings.accptance === "I Will Not Get Tech Support For This")) {
-        if($tw.settings.fileURLPrefix === '') {
-          pathRegExp = new RegExp('^/.+$');
-        } else {
-          pathRegExp = new RegExp('\/' + $tw.settings.fileURLPrefix + '\/.+$');
-        }
-      }
-      // Add the external files route handler
-      $tw.httpServer.addRoute({
-        method: "GET",
-        path: pathRegExp,
-        handler: function(request, response, state) {
-          const wikiName = findName(request.url.replace(/^\//, ''));
-          const filePrefix = $tw.settings.fileURLPrefix?$tw.settings.fileURLPrefix:'files';
-          let urlPieces = request.url.split('/');
-          // Check to make sure that the wiki name actually matches the URL
-          // Without this you could put in foo/bar/baz and get files from
-          // foo/bar if there was a wiki tehre and not on foo/bar/baz and then
-          // it would break when someone made a wiki on foo/bar/baz
-          let ok = false;
-          if(wikiName !== '' && wikiName !== 'RootWiki') {
-            ok = (request.url.replace(/^\//, '').split('/')[wikiName.split('/').length] === filePrefix);
-          } else {
-            ok = (request.url.replace(/^\//, '').split('/')[0] === filePrefix);
-          }
-          const filePath = decodeURIComponent(urlPieces.slice(urlPieces.indexOf(filePrefix)+1).join('/'));
-          const token = $tw.Bob.getCookie(request.headers.cookie, 'token');
-          const authorised = $tw.Bob.AccessCheck(wikiName, token, 'view');
-          if(authorised && ok) {
-            const basePath = $tw.ServerSide.getBasePath();
-            let pathRoot = path.resolve(basePath,$tw.settings.filePathRoot);
-            if(wikiName !== '') {
-              pathRoot = path.resolve($tw.ServerSide.getWikiPath(wikiName), 'files');
-            }
-            const pathname = path.resolve(pathRoot, filePath)
-            // Make sure that someone doesn't try to do something like ../../ to get to things they shouldn't get.
-            if(pathname.startsWith(pathRoot)) {
-              fs.exists(pathname, function(exists) {
-                if(!exists || fs.statSync(pathname).isDirectory()) {
-                  response.statusCode = 404;
-                  response.end();
-                }
-                fs.readFile(pathname, function(err, data) {
-                  if(err) {
-                    $tw.Bob.logger.error(err, {level:1})
-                    response.statusCode = 500;
-                    response.end();
-                  } else {
-                    const ext = path.parse(pathname).ext.toLowerCase();
-                    const mimeMap = $tw.settings.mimeMap || {
-                      '.aac': 'audio/aac',
-                      '.avi': 'video/x-msvideo',
-                      '.csv': 'text/csv',
-                      '.doc': 'application/msword',
-                      '.epub': 'application/epub+zip',
-                      '.gif': 'image/gif',
-                      '.html': 'text/html',
-                      '.htm': 'text/html',
-                      '.ico': 'image/x-icon',
-                      '.jpg': 'image/jpeg',
-                      '.jpeg': 'image/jpeg',
-                      '.mp3': 'audio/mpeg',
-                      '.mpeg': 'video/mpeg',
-                      '.oga': 'audio/ogg',
-                      '.ogv': 'video/ogg',
-                      '.ogx': 'application/ogg',
-                      '.png': 'image/png',
-                      '.svg': 'image/svg+xml',
-                      '.weba': 'audio/weba',
-                      '.webm': 'video/webm',
-                      '.wav': 'audio/wav'
-                    };
-                    if(mimeMap[ext] || ($tw.settings.allowUnsafeMimeTypes && $tw.settings.accptance === "I Will Not Get Tech Support For This")) {
-                      response.writeHead(200, {"Content-type": mimeMap[ext] || "text/plain"});
-                      response.end(data);
-                    } else {
-                      response.writeHead(403);
-                      response.end();
-                    }
-                  }
-                })
-              })
-            }
-          }
-        }
-      });
-    }
     // Add placeholders for other routes that load the wikis associated with
     // each route.
     $tw.httpServer.addOtherRoutes();
@@ -737,6 +283,9 @@ if($tw.node) {
 
 
   function addRoutesThing(inputObject, prefix) {
+    $tw.modules.forEachModuleOfType("wikiroute", function(title, routeDefinition) {
+      $tw.httpServer.addRoute(routeDefinition('RootWiki'));
+    });
     if(typeof inputObject === 'object') {
       Object.keys(inputObject).forEach(function (wikiName) {
         if(typeof inputObject[wikiName] === 'string') {
@@ -749,50 +298,10 @@ if($tw.node) {
             }
           }
 
-          // Make route handler
-          $tw.httpServer.addRoute({
-            method: "GET",
-            path: new RegExp('^\/' + fullName + '\/?$'),
-            handler: function(request, response, state) {
-              const token = $tw.Bob.getCookie(request.headers.cookie, 'token');
-              const authorised = $tw.Bob.AccessCheck(fullName, token, 'view');
-              let text;
-              if(authorised) {
-                // Make sure we have loaded the wiki tiddlers.
-                // This does nothing if the wiki is already loaded.
-                const exists = $tw.ServerSide.loadWiki(fullName);
-                if(exists) {
-                  // If servePlugin is not false than we strip out the filesystem
-                  // and tiddlyweb plugins if they are there and add in the
-                  // Bob plugin.
-                  const servePlugin = !$tw.settings['ws-server'].servePlugin || $tw.settings['ws-server'].servePlugin !== false;
-                  // Get the full text of the html wiki to send as the response.
-                  text = $tw.ServerSide.prepareWiki(fullName, servePlugin);
-                } else {
-                  text = "<html><p>No wiki found! Either there is no usable tiddlywiki.info file in the listed location or it isn't listed.</p></html>"
-                }
+          $tw.modules.forEachModuleOfType("wikiroute", function(title, routeDefinition) {
+            $tw.httpServer.addRoute(routeDefinition(fullName));
+          });
 
-                response.writeHead(200, {"Content-Type": state.server.get("serveType")});
-                response.end(text,"utf8");
-              }
-            }
-          });
-          // And add the favicon route for the child wikis
-          $tw.httpServer.addRoute({
-            method: "GET",
-            path: new RegExp('^\/' + fullName + '\/favicon.ico$'),
-            handler: function(request,response,state) {
-              const exists = $tw.ServerSide.loadWiki(fullName);
-              if(exists) {
-                response.writeHead(200, {"Content-Type": "image/x-icon"});
-                const buffer = $tw.Bob.Wikis[fullName].wiki.getTiddlerText('$:/favicon.ico');
-                response.end(buffer,"base64");
-              } else {
-                response.writeHead(404);
-                response.end();
-              }
-            }
-          });
           $tw.Bob.logger.log("Added route " + String(new RegExp('^\/' + fullName + '\/?$')), {level:1})
         } else {
           // recurse!
